+++ NEWS.rst
@@ -2,10 +2,10 @@
 News
 ====
 
-Latest versions of ezdxf (>= 0.9.0) are not Python 2 compatible
+Latest versions of ezdxf (>= 0.9) are not Python 2 compatible
 
-Version 0.8.10b3 - dev
-----------------------
+Version 0.8.10 - 2019-02-24
+---------------------------
 
 - BUGFIX release for Python 2 compatible main version 0.8
 - renamed groups.py into dxfgroups.py to avoid confilicts with cython